requirements.txt
Django==1.7.1
django-historicalrecords==1.1
django-reversion==1.8.5
django-simple-history==1.5.3
django-suit==0.2.12
django_select2
django-bootstrap3==4.11.0
django-rosetta==0.7.4
django-modeltranslation==0.8
django-templated-email
django-tables2
django-staticfiles
django-multiselectfield